+++ changelog.md
@@ -7,6 +7,7 @@ This update fills out the bot functionality and adds more complex features for t
 - Manage Guilds page for Guild Moderators
 - LFG Roles can be assigned to guilds
 - Manage Guild page for Admin and Guild Mods
+- /draft command for building a quick draft
 
 ### Improvements
 - Guilds can now be assigned Moderators (can add invite, assign mods, add LFG roles)